[Feature] LynxBase standalone -- Set up build project for the LynxBase module On iOS
This change introduces a dedicated build project for the iOS Base library to improve modularity.

1. Auto-generate Podspec: Adds LynxBase podspec_target and Framework subspec_target to automatically create the Pod build script for the LynxBase component.
2. Integrate with Explorer: The Lynx Base component is now compiled as part of the explorer project. This ensures that ongoing changes in other areas won't break the base component build.

issue:m-6736726571
AutoSubmit:true

+++ base/src/datauri_utils.cc
@@ -28,7 +28,7 @@ int32_t DataURIUtil::DecodeBase64(
   if (actual_buffer_size == static_cast<size_t>(-1)) {
     return 0;
   }
-  return actual_buffer_size;
+  return static_cast<int32_t>(actual_buffer_size);
 }
 
 int32_t DataURIUtil::DecodeDataURI(
